+++ stream-chat-android-compose/src/test/kotlin/io/getstream/chat/android/compose/viewmodel/channel/ChannelInfoMemberViewModelTest.kt
@@ -1,0 +1,101 @@
+/*
+ * Copyright (c) 2014-2025 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.chat.android.compose.viewmodel.channel
+
+import app.cash.turbine.test
+import io.getstream.chat.android.randomCID
+import io.getstream.chat.android.randomMember
+import io.getstream.chat.android.randomString
+import io.getstream.chat.android.ui.common.feature.channel.info.ChannelInfoMemberViewAction
+import io.getstream.chat.android.ui.common.feature.channel.info.ChannelInfoMemberViewController
+import io.getstream.chat.android.ui.common.feature.channel.info.ChannelInfoMemberViewEvent
+import io.getstream.chat.android.ui.common.state.channel.info.ChannelInfoMemberViewState
+import kotlinx.coroutines.flow.MutableSharedFlow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.test.runTest
+import org.junit.Assert.assertEquals
+import org.junit.Test
+import org.mockito.kotlin.doReturn
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.verify
+import org.mockito.kotlin.whenever
+
+internal class ChannelInfoMemberViewModelTest {
+
+    @Test
+    fun `state should reflect controller state`() = runTest {
+        val stateFlow = MutableStateFlow<ChannelInfoMemberViewState>(ChannelInfoMemberViewState.Loading)
+        val sut = Fixture()
+            .givenControllerStateFlow(stateFlow)
+            .get()
+
+        sut.state.test {
+            assertEquals(ChannelInfoMemberViewState.Loading, awaitItem())
+        }
+    }
+
+    @Test
+    fun `events should reflect controller events`() = runTest {
+        val eventsFlow = MutableSharedFlow<ChannelInfoMemberViewEvent>()
+        val sut = Fixture()
+            .givenControllerEventsFlow(eventsFlow)
+            .get()
+
+        sut.events.test {
+            val event = ChannelInfoMemberViewEvent.RemoveMember(member = randomMember())
+            eventsFlow.emit(event)
+
+            assertEquals(
+                event,
+                awaitItem(),
+            )
+        }
+    }
+
+    @Test
+    fun `onViewAction should delegate to controller`() = runTest {
+        val fixture = Fixture()
+        val sut = fixture.get()
+
+        val action = ChannelInfoMemberViewAction.BanMemberClick
+        sut.onViewAction(action)
+
+        fixture.verifyControllerOnViewAction(action)
+    }
+
+    private class Fixture {
+        private val mockController: ChannelInfoMemberViewController = mock()
+
+        fun givenControllerStateFlow(stateFlow: MutableStateFlow<ChannelInfoMemberViewState>) = apply {
+            whenever(mockController.state) doReturn stateFlow
+        }
+
+        fun givenControllerEventsFlow(eventsFlow: MutableSharedFlow<ChannelInfoMemberViewEvent>) = apply {
+            whenever(mockController.events) doReturn eventsFlow
+        }
+
+        fun verifyControllerOnViewAction(action: ChannelInfoMemberViewAction) = apply {
+            verify(mockController).onViewAction(action)
+        }
+
+        fun get() = ChannelInfoMemberViewModel(
+            cid = randomCID(),
+            memberId = randomString(),
+            controllerProvider = { mockController },
+        )
+    }
+}