+++ CHANGELOG.md
@@ -70,10 +70,12 @@
 ### 🐞 Fixed
 
 ### ⬆️ Improved
+- Remove the fixed height modifier in `FileUploadItem`. [#5932](https://github.com/GetStream/stream-chat-android/pull/5932)
 
 ### ✅ Added
 
 ### ⚠️ Changed
+- Deprecate unused component `AudioWaveVSeekbar`. [#5932](https://github.com/GetStream/stream-chat-android/pull/5932)
 
 ### ❌ Removed